langchain-edenmarco-udemy-ws/langchain-mcp/pyproject.toml
[project]
name = "langchain-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.1.0",
    "fastmcp>=2.14.3",
    "isort>=7.0.0",
    "langchain>=1.2.6",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "langchain-mcp-adapters>=0.2.1",
    "python-dotenv>=1.2.1",
]
